Move SQL parsing functions to sql. module
